Switch to a "Save" button
Needs work on the dialogs.
